requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==21.4.0
colorama==0.4.5
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
factory-boy==3.2.1
Faker==13.15.0
inflection==0.5.1
iniconfig==1.1.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
